1.完善Locomotion2.修复部分bug

+++ Assets/Scripts/Code/Module/Config/SceneGroup/ConfigSceneGroupAction/ConfigSceneGroupAddVariableAction.cs
@@ -39,7 +39,7 @@ namespace TaoTie
                 val += flag;
                 if (Limit)
                 {
-                    val = (int) Mathf.Clamp(val, MinValue, MaxValue);
+                    val = Mathf.Clamp(val, MinValue, MaxValue);
                 }
 
                 aimSceneGroup.Variable.Set(Key, val);
